--- src/main/java/org/factoriaf5/happypaws/SecurityConfig.java
@@ -1,27 +1,0 @@
-package org.factoriaf5.happypaws;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-            .csrf(csrf -> csrf.disable()) // Desactivamos CSRF
-            .authorizeHttpRequests(auth -> auth
-                .anyRequest().permitAll() // Permitimos todas las rutas
-            );
-        return http.build();
-    }
-}
