+++ src/main/java/org/factoriaf5/happypaws/register/RegisterService.java
@@ -1,0 +1,5 @@
+package org.factoriaf5.happypaws.register;
+
+public interface RegisterService {
+    RegisterDTOResponse registerUser(RegisterDTORequest dto);
+}
