--- src/main/java/org/factoriaf5/happypaws/register/services/RegisterService.java
@@ -1,8 +1,0 @@
-package org.factoriaf5.happypaws.register.services;
-
-import org.factoriaf5.happypaws.register.dtos.RegisterDTORequest;
-import org.factoriaf5.happypaws.register.dtos.RegisterDTOResponse;
-
-public interface RegisterService {
-    RegisterDTOResponse registerUser(RegisterDTORequest dto);
-}
