+++ src/main/java/org/factoriaf5/happypaws/role/RoleRepository.java
@@ -1,7 +1,9 @@
 package org.factoriaf5.happypaws.role;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
     
+    public Optional<RoleEntity> findByName(String name);
 }
