--- src/main/java/org/factoriaf5/happypaws/temp/Role.java
@@ -1,27 +1,0 @@
-package org.factoriaf5.happypaws.temp;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "roles")
-public class Role {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    public Role() {}
-
-    public Role(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() { return id; }
-    public String getName() { return name; }
-
-    public void setId(Long id) { this.id = id; }
-    public void setName(String name) { this.name = name; }
-}
