--- src/main/java/org/factoriaf5/happypaws/temp/RoleRepository.java
@@ -1,11 +1,0 @@
-package org.factoriaf5.happypaws.temp;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findByName(String name);
-}
