--- src/main/java/org/factoriaf5/happypaws/temp/User.java
@@ -1,56 +1,0 @@
-package org.factoriaf5.happypaws.temp;
-
-import jakarta.persistence.*;
-import java.util.Set;
-
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;
-    private String password;
-    private String fullName;
-    private String phone;
-    private String email;
-
-    @ManyToMany(fetch = FetchType.EAGER)
-    @JoinTable(
-            name = "user_roles",
-            joinColumns = @JoinColumn(name = "user_id"),
-            inverseJoinColumns = @JoinColumn(name = "role_id")
-    )
-    private Set<Role> roles;
-
-    public User() {}
-
-    public User(Long id, String username, String password, String fullName, String phone, String email, Set<Role> roles) {
-        this.id = id;
-        this.username = username;
-        this.password = password;
-        this.fullName = fullName;
-        this.phone = phone;
-        this.email = email;
-        this.roles = roles;
-    }
-
-    // Getters & setters (solo lo necesario)
-    public Long getId() { return id; }
-    public String getUsername() { return username; }
-    public String getPassword() { return password; }
-    public String getFullName() { return fullName; }
-    public String getPhone() { return phone; }
-    public String getEmail() { return email; }
-    public Set<Role> getRoles() { return roles; }
-
-    public void setId(Long id) { this.id = id; }
-    public void setUsername(String username) { this.username = username; }
-    public void setPassword(String password) { this.password = password; }
-    public void setFullName(String fullName) { this.fullName = fullName; }
-    public void setPhone(String phone) { this.phone = phone; }
-    public void setEmail(String email) { this.email = email; }
-    public void setRoles(Set<Role> roles) { this.roles = roles; }
-}
