--- src/main/java/org/factoriaf5/happypaws/temp/UserRepository.java
@@ -1,12 +1,0 @@
-package org.factoriaf5.happypaws.temp;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-    boolean existsByUsername(String username);
-}
