--- src/main/java/org/factoriaf5/happypaws/user/controllers/UserController.java
@@ -1,5 +1,0 @@
-package org.factoriaf5.happypaws.user.controllers;
-
-public class UserController {
-    
-}
