--- src/main/java/org/factoriaf5/happypaws/user/exceptions/UserNotFoundException.java
@@ -1,5 +1,0 @@
-package org.factoriaf5.happypaws.user.exceptions;
-
-public class UserNotFoundException {
-    
-}
