--- src/main/java/org/factoriaf5/happypaws/user/mappers/UserMapper.java
@@ -1,5 +1,0 @@
-package org.factoriaf5.happypaws.user.mappers;
-
-public class UserMapper {
-    
-}
