--- src/main/java/org/factoriaf5/happypaws/user/repositories/UserRepository.java
@@ -1,5 +1,0 @@
-package org.factoriaf5.happypaws.user.repositories;
-
-public class UserRepository {
-    
-}
