--- src/main/java/org/factoriaf5/happypaws/user/services/UserService.java
@@ -1,5 +1,0 @@
-package org.factoriaf5.happypaws.user.services;
-
-public class UserService {
-    
-}
