--- src/main/java/org/factoriaf5/happypaws/user/validators/UserValidator.java
@@ -1,5 +1,0 @@
-package org.factoriaf5.happypaws.user.validators;
-
-public class UserValidator {
-    
-}
