+++ src/test/java/org/factoriaf5/happypaws/register/RegisterServiceImplTest.java
@@ -1,14 +1,9 @@
 package org.factoriaf5.happypaws.register;
 
-import org.factoriaf5.happypaws.register.dtos.RegisterDTORequest;
-import org.factoriaf5.happypaws.register.dtos.RegisterDTOResponse;
-import org.factoriaf5.happypaws.register.exceptions.UserAlreadyExistsException;
-import org.factoriaf5.happypaws.register.services.RegisterServiceImpl;
-import org.factoriaf5.happypaws.register.validators.RegisterValidator;
-import org.factoriaf5.happypaws.temp.Role;
-import org.factoriaf5.happypaws.temp.RoleRepository;
-import org.factoriaf5.happypaws.temp.User;
-import org.factoriaf5.happypaws.temp.UserRepository;
+import org.factoriaf5.happypaws.role.RoleEntity;
+import org.factoriaf5.happypaws.role.RoleRepository;
+import org.factoriaf5.happypaws.user.UserEntity;
+import org.factoriaf5.happypaws.user.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -42,9 +37,9 @@ class RegisterServiceImplTest {
                 "12345678A", "Juan Perez", "600123456", "juan@mail.com", "pass123", "pass123"
         );
 
-        Role clientRole = new Role(1L, "CLIENT");
+        RoleEntity userRole = new RoleEntity();
         Mockito.when(userRepository.findByUsername(dto.username())).thenReturn(Optional.empty());
-        Mockito.when(roleRepository.findByName("CLIENT")).thenReturn(Optional.of(clientRole));
+        Mockito.when(roleRepository.findByName("CLIENT")).thenReturn(Optional.of(userRole));
         Mockito.when(passwordEncoder.encode(dto.password())).thenReturn("hashedPass");
 
         RegisterDTOResponse response = service.registerUser(dto);
@@ -53,7 +48,7 @@ class RegisterServiceImplTest {
         assertEquals(dto.fullName(), response.fullName());
         assertEquals(dto.phone(), response.phone());
         assertEquals(dto.email(), response.email());
-        Mockito.verify(userRepository).save(any(User.class));
+        Mockito.verify(userRepository).save(any(UserEntity.class));
     }
 
     @Test
@@ -61,7 +56,7 @@ class RegisterServiceImplTest {
         RegisterDTORequest dto = new RegisterDTORequest(
                 "12345678A", "Juan Perez", "600123456", "juan@mail.com", "pass123", "pass123"
         );
-        Mockito.when(userRepository.findByUsername(dto.username())).thenReturn(Optional.of(new User()));
+        Mockito.when(userRepository.findByUsername(dto.username())).thenReturn(Optional.of(new UserEntity()));
         assertThrows(UserAlreadyExistsException.class, () -> service.registerUser(dto));
     }
 }
